New folder system, cleaner code
So I basically moved Javascript-Practice repo stuff into this new
website. Its cleaner, I seperated all of the javascript and the css from
the html. I added a timer and I commented the entire code, so that you
guys can understand it more easily.
I'll be sure to keep updating this one.