Apere System for Food Stuffs
So we're porting the apere system we made for the library to the web. It's going to be remade from the ground up to hopefully be more scalable this time. So html/css and javascript will be the bread and butter. Using local server to begin with, but I'm thinking of running it through azure somewhere down the line.